Splitting the Shared and UUtils files in components
