+++ src/Platform/App/components/Map/ClusterTransportationRoutesLayer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { GeoJSON } from 'react-leaflet';
 import { FeatureCollection } from 'geojson';
 import { MapCoordinates, YearlyResult } from '../../models/Types';
-import { serviceUrl } from '../Shared/config';
+import { serviceUrl } from '../Utils/config';
 
 interface Props {
   facilityCoordinates: MapCoordinates;
