+++ src/Platform/App/components/Map/MapContainer.tsx
@@ -52,8 +52,8 @@ import { trackEvent } from '../../../Utils/gaAnalytics';
 import { InputContainer } from '../Inputs/InputContainer';
 import { checkFrcsValidity, checkTeaValidity } from '../Inputs/validation';
 import { ResultsContainer } from '../Results/ResultsContainer';
-import { serviceUrl } from '../Shared/config';
-import { convertGeoJSON } from '../Shared/util';
+import { serviceUrl } from '../Utils/config';
+import { convertGeoJSON } from '../Utils/util';
 import { ClusterTransportationMoveInLayer } from './ClusterTransportationMoveInLayer';
 import { ClusterTransportationRoutesLayer } from './ClusterTransportationRoutesLayer';
 import { CustomMarker } from './CustomMarker';
