+++ src/Platform/App/components/Results/Frcs/FrcsSkippedClusterResultsContainer.tsx
@@ -6,7 +6,7 @@ import {
 } from '../../../models/Types';
 import { Button } from 'reactstrap';
 import { ReactTable } from '../../Shared/ReactTable';
-import { formatNumber } from '../../Shared/util';
+import { formatNumber } from '../../Utils/util';
 import { Cell } from 'react-table';
 
 interface Props {
