+++ src/Platform/App/components/Results/Frcs/FrcsTables.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatNumber } from '../../Shared/util';
+import { formatNumber } from '../../Utils/util';
 import { YearlyResult } from '../../../models/Types';
 import { Table } from 'reactstrap';
