+++ src/Platform/App/components/Results/LCA/LCAResults.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatNumber } from '../../Shared/util';
+import { formatNumber } from '../../Utils/util';
 import { LCAresults } from '../../../models/Types';
 
 interface Props {
