+++ src/Platform/App/components/Results/ResultsTables.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatCurrency, formatNumber } from '../Shared/util';
+import { formatCurrency, formatNumber } from '../Utils/util';
 import {
   AllYearsResults,
   FrcsInputs,
