+++ src/Platform/App/components/Results/Technoeconomic/DetailedResults/ConstantLAC.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatNumber } from '../../../Shared/util';
+import { formatNumber } from '../../../Utils/util';
 import { ConstantLevelAnnualCostMod } from '@ucdavis/tea/output.model';
 
 interface Props {
