+++ src/Platform/App/components/Results/Technoeconomic/DetailedResults/ExpensesBaseYear.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatNumber } from '../../../Shared/util';
+import { formatNumber } from '../../../Utils/util';
 import {
   ExpensesBaseYearModGPO,
   ExpensesBaseYearModGP
