+++ src/Platform/App/components/Results/Technoeconomic/DetailedResults/Financing.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { formatNumber, formatCurrency } from '../../../Shared/util';
+import { formatNumber, formatCurrency } from '../../../Utils/util';
 import { FinancingInputMod } from '@ucdavis/tea/input.model';
 import { FinancingMod } from '@ucdavis/tea/output.model';
