+++ src/Platform/App/components/Utils/myBase64Image.ts
