+++ src/Platform/App/components/Utils/react-table-config.d.ts
