+++ src/Platform/App/components/Utils/util.ts
@@ -1,5 +1,5 @@
+import { Feature } from 'geojson';
 import { ClusterResult } from '../../models/Types';
-import { Feature, FeatureCollection } from 'geojson';
 
 export const formatNumber = (num: number, digits?: number) => {
   if (!num) {
